+++ resources/components/laraknife/icons/delete-file.blade.php
@@ -1,0 +1,2 @@
+@props(['url' => '', 'file' => ''])
+<a href="{{$url}}/{{$file}}"><x-laraknife.icons.delete class="text-primary"/></a>