+++ resources/components/laraknife/panels/simple.blade.php
@@ -1,0 +1,4 @@
+@props(['legend' => ''])
+<fieldset class="lkn-panel">
+     {{$slot}}
+</fieldset>